+++ .idea/deploymentTargetDropDown.xml
@@ -7,11 +7,11 @@
         <deviceKey>
           <Key>
             <type value="VIRTUAL_DEVICE_PATH" />
-            <value value="C:\Users\Juan Camilo\.android\avd\Galaxy_Nexus_API_23.avd" />
+            <value value="C:\Users\Juan Camilo\.android\avd\Galaxy_Nexus_API_22.avd" />
           </Key>
         </deviceKey>
       </Target>
     </targetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2021-11-08T02:48:24.817342600Z" />
+    <timeTargetWasSelectedWithDropDown value="2021-11-07T02:28:52.962090300Z" />
   </component>
 </project>
