+++ .idea/misc.xml
@@ -5,15 +5,12 @@
       <map>
         <entry key="..\:/Kotlin_Projects/Vinilos/app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.3567708333333333" />
         <entry key="..\:/Kotlin_Projects/Vinilos/app/src/main/res/drawable/ic_launcher_background.xml" value="0.3567708333333333" />
-        <entry key="..\:/Kotlin_Projects/Vinilos/app/src/main/res/layout/activity_album.xml" value="0.2" />
+        <entry key="..\:/Kotlin_Projects/Vinilos/app/src/main/res/layout/activity_album.xml" value="0.2740036231884058" />
         <entry key="..\:/Kotlin_Projects/Vinilos/app/src/main/res/layout/activity_main.xml" value="0.25" />
-        <entry key="..\:/Kotlin_Projects/Vinilos/app/src/main/res/layout/activity_musico.xml" value="0.1" />
         <entry key="..\:/Kotlin_Projects/Vinilos/app/src/main/res/layout/detalle_album.xml" value="0.5" />
         <entry key="..\:/Kotlin_Projects/Vinilos/app/src/main/res/layout/fragment_album.xml" value="0.20885416666666667" />
         <entry key="..\:/Kotlin_Projects/Vinilos/app/src/main/res/layout/fragment_detalle_album.xml" value="0.20885416666666667" />
-        <entry key="..\:/Kotlin_Projects/Vinilos/app/src/main/res/layout/fragment_musico.xml" value="0.2644927536231884" />
         <entry key="..\:/Kotlin_Projects/Vinilos/app/src/main/res/layout/item_album.xml" value="0.2740036231884058" />
-        <entry key="..\:/Kotlin_Projects/Vinilos/app/src/main/res/layout/item_musico.xml" value="0.2644927536231884" />
       </map>
     </option>
   </component>
