--- app/src/androidTest/java/co/edu/uniandes/fourbidden/vinilos/vista/PruebaListarMusicos.kt
@@ -1,79 +1,0 @@
-package co.edu.uniandes.fourbidden.vinilos.vista
-
-
-import android.view.View
-import android.view.ViewGroup
-import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.action.ViewActions.click
-import androidx.test.espresso.matcher.ViewMatchers.*
-import androidx.test.filters.LargeTest
-import androidx.test.rule.ActivityTestRule
-import androidx.test.runner.AndroidJUnit4
-import co.edu.uniandes.fourbidden.vinilos.R
-import org.hamcrest.Description
-import org.hamcrest.Matcher
-import org.hamcrest.Matchers.allOf
-import org.hamcrest.TypeSafeMatcher
-import org.junit.Rule
-import org.junit.Test
-import org.junit.runner.RunWith
-
-
-@RunWith(AndroidJUnit4::class)
-class PruebaListarMusicos {
-
-    @get:Rule
-    val mActivityTestRule = ActivityTestRule(MainActivity::class.java)
-
-    @Test
-    fun pruebaListarMusicos() {
-        val materialButton = onView(
-            allOf(
-                withId(R.id.bingreso), withText("Entrar"),
-                childAtPosition(
-                    childAtPosition(
-                        withId(android.R.id.content),
-                        0
-                    ),
-                    4
-                ),
-                isDisplayed()
-            )
-        )
-        materialButton.perform(click())
-        Thread.sleep(5000)
-        val materialButton2 = onView(
-            allOf(
-                withId(R.id.btMusico), withText("Músicos"),
-                childAtPosition(
-                    childAtPosition(
-                        withId(android.R.id.content),
-                        0
-                    ),
-                    4
-                ),
-                isDisplayed()
-            )
-        )
-        materialButton2.perform(click())
-        Thread.sleep(5000)
-    }
-
-    private fun childAtPosition(
-        parentMatcher: Matcher<View>, position: Int
-    ): Matcher<View> {
-
-        return object : TypeSafeMatcher<View>() {
-            override fun describeTo(description: Description) {
-                description.appendText("Child at position $position in parent ")
-                parentMatcher.describeTo(description)
-            }
-
-            public override fun matchesSafely(view: View): Boolean {
-                val parent = view.parent
-                return parent is ViewGroup && parentMatcher.matches(parent)
-                        && view == parent.getChildAt(position)
-            }
-        }
-    }
-}
