+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
     package="co.edu.uniandes.fourbidden.vinilos">
 
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
@@ -26,14 +25,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
-        <activity
-            android:name=".vista.MusicoActivity"
-            android:exported="false" />
-        <activity
-            android:name=".vista.ColeccionistaActivity"
-            android:exported="false"
-            tools:ignore="MissingClass" />
     </application>
 
 </manifest>
