--- app/src/main/java/co/edu/uniandes/fourbidden/vinilos/modelo/Coleccionista.kt
@@ -1,9 +1,0 @@
-package co.edu.uniandes.fourbidden.vinilos.modelo
-
-data class Coleccionista (
-    val id: Int,
-    val name: String,
-    val email: String,
-    //val comments: List,
-    //val collectorAlbums: List
-    )
