+++ app/src/main/java/co/edu/uniandes/fourbidden/vinilos/modelo/servicio/ServiceAdapter.kt
@@ -16,9 +16,6 @@ import org.json.JSONArray
 import org.json.JSONObject
 import java.time.LocalDate
 import java.time.LocalDate.parse
-import kotlin.coroutines.resume
-import kotlin.coroutines.resumeWithException
-import kotlin.coroutines.suspendCoroutine
 
 @Suppress("RedundantSamConstructor")
 class ServiceAdapter constructor(context: Context) {
@@ -35,10 +32,8 @@ class ServiceAdapter constructor(context: Context) {
     private val requestQueue: RequestQueue by lazy {
         Volley.newRequestQueue(context.applicationContext)
     }
-
-
-
-    suspend fun getAlbums() = suspendCoroutine<List<Album>>{ cont->
+    @RequiresApi(Build.VERSION_CODES.O)
+    fun getAlbums(onComplete:(resp:List<Album>)->Unit, onError: (error: VolleyError)->Unit){
         requestQueue.add(getRequest("albums",
             Response.Listener<String> { response ->
                 val resp = JSONArray(response)
@@ -46,17 +41,20 @@ class ServiceAdapter constructor(context: Context) {
                 for (i in 0 until resp.length()) {
                     val item = resp.getJSONObject(i)
                     val fecha : String =  item!!.getString("releaseDate").substringBefore(delimiter = "T", missingDelimiterValue = "2000-01-01")
+
                     var releaseDate : LocalDate = parse(fecha)
                     val listTrack = mutableListOf<Track>()
+
+
+
                     list.add(i, Album(id = item.getString("id"),name = item.getString("name"), cover = item.getString("cover"), recordLabel = item.getString("recordLabel"), releaseDate = releaseDate, genre = item.getString("genre"), description = item.getString("description"), tracks = listTrack))
                 }
-                cont.resume(list)
+                onComplete(list)
             },
             Response.ErrorListener {
-                cont.resumeWithException(it)
+                onError(it)
             }))
     }
-
 
 
     @RequiresApi(Build.VERSION_CODES.O)
