+++ app/src/main/java/co/edu/uniandes/fourbidden/vinilos/vista/adapter/AlbumsAdapter.kt
@@ -4,7 +4,6 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import android.widget.ImageView
 import androidx.annotation.LayoutRes
-import androidx.core.net.toUri
 import androidx.databinding.DataBindingUtil
 import androidx.navigation.findNavController
 import androidx.recyclerview.widget.RecyclerView
@@ -12,9 +11,6 @@ import co.edu.uniandes.fourbidden.vinilos.R
 import co.edu.uniandes.fourbidden.vinilos.databinding.ItemAlbumBinding
 import co.edu.uniandes.fourbidden.vinilos.modelo.Album
 import co.edu.uniandes.fourbidden.vinilos.vista.fragmentos.AlbumFragmentDirections
-import com.bumptech.glide.Glide
-import com.bumptech.glide.load.engine.DiskCacheStrategy
-import com.bumptech.glide.request.RequestOptions
 import com.squareup.picasso.Picasso
 
 class AlbumsAdapter: RecyclerView.Adapter<AlbumsAdapter.AlbumViewHolder>() {
@@ -32,21 +28,6 @@ class AlbumsAdapter: RecyclerView.Adapter<AlbumsAdapter.AlbumViewHolder>() {
         return AlbumViewHolder(withDataBinding)
     }
 
-
-    override fun onBindViewHolder(holder: AlbumViewHolder, position: Int) {
-        holder.viewDataBinding.also {
-            it.album = albums[position]
-        }
-        holder.bind(albums[position])
-        holder.viewDataBinding.root.setOnClickListener {
-            val action = AlbumFragmentDirections.actionAlbumFragmentToFragmentDetalleAlbum(albums[position].id)
-            // Navigate using that action
-            holder.viewDataBinding.root.findNavController().navigate(action)
-        }
-    }
-
-
-    /*
     override fun onBindViewHolder(holder: AlbumViewHolder, position: Int) {
 
         holder.viewDataBinding.also {
@@ -60,24 +41,12 @@ class AlbumsAdapter: RecyclerView.Adapter<AlbumsAdapter.AlbumViewHolder>() {
             holder.viewDataBinding.root.findNavController().navigate(action)
         }
     }
-*/
+
     override fun getItemCount(): Int {
         return albums.size
     }
     class AlbumViewHolder(val viewDataBinding: ItemAlbumBinding) :
         RecyclerView.ViewHolder(viewDataBinding.root) {
-
-
-       fun bind(album: Album) {
-            Glide.with(itemView)
-                .load(album.cover.toUri().buildUpon().scheme("https").build())
-                .apply(RequestOptions()
-                    //.placeholder(R.drawable.loading_animation)
-                    .diskCacheStrategy(DiskCacheStrategy.ALL)
-                    .error(R.drawable.ic_broken_image))
-                .into(viewDataBinding.albumCover)
-        }
-
 
         companion object {
             @LayoutRes
