--- app/src/main/java/co/edu/uniandes/fourbidden/vinilos/vista/fragmentos/ColeccionistaFragment.kt
@@ -1,80 +1,0 @@
-package co.edu.uniandes.fourbidden.vinilos.vista.fragmentos
-
-import android.os.Build
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.Toast
-import androidx.annotation.RequiresApi
-import androidx.fragment.app.Fragment
-import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProvider
-import androidx.recyclerview.widget.GridLayoutManager
-import androidx.recyclerview.widget.RecyclerView
-import co.edu.uniandes.fourbidden.vinilos.databinding.FragmentColeccionistaBinding
-import co.edu.uniandes.fourbidden.vinilos.modelo.Coleccionista
-import co.edu.uniandes.fourbidden.vinilos.vista.adapter.ColeccionistaAdapter
-import co.edu.uniandes.fourbidden.vinilos.vistamodelo.ColeccionistaViewModel
-
-/**
- * Objeto para visualizar los musicos.
- */
-class ColeccionistaFragment : Fragment() {
-
-    private var _binding: FragmentColeccionistaBinding? = null
-    private val binding get() = _binding!!
-    private lateinit var recyclerView: RecyclerView
-    private lateinit var viewModel: ColeccionistaViewModel
-    private var viewModelAdapter: ColeccionistaAdapter? = null
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = FragmentColeccionistaBinding.inflate(inflater, container, false)
-        val view = binding.root
-        viewModelAdapter = ColeccionistaAdapter()
-        return view
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        recyclerView = binding.coleccionistaRv
-        recyclerView.layoutManager = GridLayoutManager(context, 3)//LinearLayoutManager(context)
-        recyclerView.layoutManager
-        recyclerView.adapter = viewModelAdapter
-    }
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    override fun onActivityCreated(savedInstanceState: Bundle?) {
-        super.onActivityCreated(savedInstanceState)
-        val activity = requireNotNull(this.activity) {
-            "You can only access the viewModel after onActivityCreated()"
-        }
-
-        viewModel = ViewModelProvider(this, ColeccionistaViewModel.Factory(activity.application)).get(
-            ColeccionistaViewModel::class.java)
-
-        viewModel.coleccionistas.observe(viewLifecycleOwner, Observer<List<Coleccionista>> {
-
-            it.apply {
-                viewModelAdapter!!.coleccionistas = this
-            }
-        })
-        viewModel.eventNetworkError.observe(viewLifecycleOwner, Observer<Boolean> { isNetworkError ->
-            if (isNetworkError) onNetworkError()
-        })
-    }
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-
-    @RequiresApi(Build.VERSION_CODES.O)
-    private fun onNetworkError() {
-        if(!viewModel.isNetworkErrorShown.value!!) {
-            Toast.makeText(activity, "Network Error", Toast.LENGTH_LONG).show()
-            viewModel.onNetworkErrorShown()
-        }
-    }
-}
