--- app/src/main/res/layout/album_musico.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-        <variable name="album" type="co.edu.uniandes.fourbidden.vinilos.modelo.Album"/>
-    </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-
-
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="1dp"
-        android:layout_marginBottom="1dp"
-        android:background="@android:color/transparent"
-        android:backgroundTint="@android:color/transparent">
-
-        <TextView
-            android:id="@+id/album_name"
-            style="@style/TextAppearance.AppCompat.Large"
-            android:layout_width="250dp"
-            android:layout_height="30dp"
-            android:layout_marginStart="8dp"
-            android:text="@{album.name}"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
