+++ app/src/main/res/layout/item_album.xml
@@ -17,7 +17,7 @@
             android:layout_height="100dp"
             android:layout_marginTop="25dp"
             app:cardBackgroundColor="@android:color/transparent"
-            app:cardCornerRadius="1dp"
+            app:cardCornerRadius="6dp"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent">
