--- app/src/main/res/layout/item_coleccionista.xml
@@ -1,60 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-    <layout xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools">
-
-        <data>
-            <variable name="coleccionista" type="co.edu.uniandes.fourbidden.vinilos.modelo.Coleccionista"/>
-        </data>
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="119dp"
-        android:layout_height="133dp">
-
-        <TextView
-            android:id="@+id/musico_name"
-            style="@style/textoLista"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_gravity="bottom"
-            android:layout_marginStart="5dp"
-            android:layout_marginEnd="5dp"
-            android:background="@android:color/transparent"
-            android:maxLines="1"
-            android:text="@{coleccionista.name}"
-            android:textAlignment="center"
-            android:textColor="@color/black"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.0"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/cardView" />
-
-        <androidx.cardview.widget.CardView
-            android:id="@+id/cardView"
-            android:layout_width="81dp"
-            android:layout_height="83dp"
-            android:layout_marginStart="20dp"
-            android:layout_marginTop="16dp"
-            app:cardBackgroundColor="@android:color/transparent"
-            app:cardCornerRadius="1dp"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent">
-
-            <ImageView
-                android:id="@+id/album_cover"
-                android:layout_width="80dp"
-                android:layout_height="83dp"
-                android:scaleType="fitXY"
-                app:layout_constraintBottom_toBottomOf="parent"
-                app:layout_constraintEnd_toEndOf="parent"
-                app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
-                app:srcCompat="@drawable/coleccionista" />
-        </androidx.cardview.widget.CardView>
-
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-
-
-    </layout>
