--- app/src/main/res/navigation/nav_graph_musico.xml
@@ -1,27 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/nav_graph_musico"
-    app:startDestination="@id/musicoFragment">
-
-    <fragment
-        android:id="@+id/musicoFragment"
-        android:name="co.edu.uniandes.fourbidden.vinilos.vista.fragmentos.MusicoFragment"
-        android:label="@string/titulolistamusico"
-        tools:layout="@layout/fragment_musico" >
-        <action
-            android:id="@+id/action_musicoFragment_to_fragmentDetalleMusico"
-            app:destination="@id/fragmentDetalleMusico" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/fragmentDetalleMusico"
-        android:name="co.edu.uniandes.fourbidden.vinilos.vista.fragmentos.DetalleMusicoFragment"
-        android:label="@string/titulodetallemusico"
-        tools:layout="@layout/detalle_musico" >
-        <argument
-            android:name="musicoId"
-            app:argType="string" />
-    </fragment>
-</navigation>
