+++ app/src/main/res/values/styles.xml
@@ -23,27 +23,8 @@
         <item name="android:textColor">#040404</item>
         <item name="android:fontFamily">@font/comfortaa</item>
         <item name="android:textSize">56dp</item>
-        <item name="android:background">#32FCFD</item>
-    </style>
-
-    <style name="customButton" parent="Widget.MaterialComponents.Button">
-        <item name="android:textColor">#F8F5F5</item>
-        <item name="android:fontFamily">@font/comfortaa</item>
-        <item name="android:textSize">11dp</item>
-        <item name="strokeWidth">3dp</item>
-        <item name="strokeColor">#32FCFD</item>
-        <item name="android:background">#32FCFD</item>
+        <item name="android:background">#32FCFDFD</item>
 
     </style>
-    <style name="botonMenu" parent="Widget.MaterialComponents.Button">
-        <item name="android:layout_height">wrap_content</item>
-        <item name="android:layout_width">120dp</item>
-        <item name="strokeColor">@color/secondaryColor</item>
-        <item name="strokeWidth">3dp</item>
-        <item name="android:fontFamily">@font/comfortaa</item>
-        <item name="android:textSize">11dp</item>
-        <item name="android:textColor">@color/white</item>
-    </style>
-
 
 </resources>
