+++ app/src/test/java/co/edu/uniandes/fourbidden/vinilos/ExampleUnitTest.kt
@@ -1,7 +1,5 @@
 package co.edu.uniandes.fourbidden.vinilos
 
-import android.app.Application
-import co.edu.uniandes.fourbidden.vinilos.modelo.servicio.ServiceAdapterMusico
 import org.junit.Test
 
 import org.junit.Assert.*
@@ -14,7 +12,6 @@ import org.junit.Assert.*
 class ExampleUnitTest {
     @Test
     fun addition_isCorrect() {
-
         assertEquals(4, 2 + 2)
     }
 }
